requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
fastapi==0.75.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
